Refactor: arrange helpers to helper_test.go
Many helpers are spread over e2e directory. Now I extract such helpers
and arrange them to helper_test.go.
