+++ e2e/e2e_test.go
@@ -1,7 +1,6 @@
 package e2e
 
 import (
-	"bytes"
 	_ "embed"
 	"encoding/json"
 	"errors"
@@ -915,41 +914,4 @@ func checkLVIsDeletedInLVM(volName string) error {
 		return fmt.Errorf("target LV exists %s", volName)
 	}
 	return nil
-}
-
-func countLVMs() (int, error) {
-	stdout, err := exec.Command("sudo", "lvs", "-o", "lv_name", "--noheadings").Output()
-	if err != nil {
-		return -1, fmt.Errorf("failed to lvs. stdout %s, err %v", stdout, err)
-	}
-	return bytes.Count(stdout, []byte("\n")), nil
-}
-
-func getNodeAnnotationMapWithPrefix(prefix string) (map[string]map[string]string, error) {
-	stdout, stderr, err := kubectl("get", "node", "-o", "json")
-	if err != nil {
-		return nil, fmt.Errorf("stdout=%sr stderr=%s, err=%v", stdout, stderr, err)
-	}
-
-	var nodeList corev1.NodeList
-	err = json.Unmarshal(stdout, &nodeList)
-	if err != nil {
-		return nil, err
-	}
-
-	capacities := make(map[string]map[string]string)
-	for _, node := range nodeList.Items {
-		if node.Name == "topolvm-e2e-control-plane" {
-			continue
-		}
-
-		capacities[node.Name] = make(map[string]string)
-		for k, v := range node.Annotations {
-			if !strings.HasPrefix(k, prefix) {
-				continue
-			}
-			capacities[node.Name][k] = v
-		}
-	}
-	return capacities, nil
 }
