+++ e2e/helper_test.go
@@ -1,16 +1,21 @@
+// This provides not test itself but helpers.
+
 package e2e
 
 import (
 	"bytes"
+	"encoding/json"
+	"errors"
 	"fmt"
 	"os/exec"
 	"path/filepath"
+	"strings"
 
 	"github.com/google/go-cmp/cmp"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-
 	"github.com/topolvm/topolvm"
+	corev1 "k8s.io/api/core/v1"
 )
 
 type CleanupContext struct {
@@ -47,6 +52,43 @@ func containString(s []string, target string) bool {
 		}
 	}
 	return false
+}
+
+func countLVMs() (int, error) {
+	stdout, err := exec.Command("sudo", "lvs", "-o", "lv_name", "--noheadings").Output()
+	if err != nil {
+		return -1, fmt.Errorf("failed to lvs. stdout %s, err %v", stdout, err)
+	}
+	return bytes.Count(stdout, []byte("\n")), nil
+}
+
+func getNodeAnnotationMapWithPrefix(prefix string) (map[string]map[string]string, error) {
+	stdout, stderr, err := kubectl("get", "node", "-o", "json")
+	if err != nil {
+		return nil, fmt.Errorf("stdout=%sr stderr=%s, err=%v", stdout, stderr, err)
+	}
+
+	var nodeList corev1.NodeList
+	err = json.Unmarshal(stdout, &nodeList)
+	if err != nil {
+		return nil, err
+	}
+
+	capacities := make(map[string]map[string]string)
+	for _, node := range nodeList.Items {
+		if node.Name == "topolvm-e2e-control-plane" {
+			continue
+		}
+
+		capacities[node.Name] = make(map[string]string)
+		for k, v := range node.Annotations {
+			if !strings.HasPrefix(k, prefix) {
+				continue
+			}
+			capacities[node.Name][k] = v
+		}
+	}
+	return capacities, nil
 }
 
 func commonBeforeEach() CleanupContext {
@@ -88,4 +130,30 @@ func commonAfterEach(cc CleanupContext) {
 			return nil
 		}).Should(Succeed())
 	}
+}
+
+type lvinfo struct {
+	lvPath string
+	size   string
+	vgName string
+}
+
+func getLVInfo(lvName string) (*lvinfo, error) {
+	stdout, err := exec.Command("sudo", "lvdisplay", "-c", "--select", "lv_name="+lvName).Output()
+	if err != nil {
+		return nil, fmt.Errorf("err=%v, stdout=%s", err, stdout)
+	}
+	output := strings.TrimSpace(string(stdout))
+	if output == "" {
+		return nil, fmt.Errorf("lv_name ( %s ) not found", lvName)
+	}
+	lines := strings.Split(output, "\n")
+	if len(lines) != 1 {
+		return nil, errors.New("found multiple lvs")
+	}
+	items := strings.Split(strings.TrimSpace(lines[0]), ":")
+	if len(items) < 4 {
+		return nil, fmt.Errorf("invalid format: %s", lines[0])
+	}
+	return &lvinfo{lvPath: items[0], vgName: items[1], size: items[3]}, nil
 }
